+++ tests/testthat/setup.R
@@ -1,4 +1,4 @@
 pre_test_options <- options(
-  readr.show_types = FALSE,
+  readr.show_col_types = FALSE,
   readr.show_progress = FALSE
 )