+++ NEWS.md
@@ -1,5 +1,7 @@
 # readr (development version)
 
+* `options(readr.show_col_types = FALSE)` now works as intended (#1250)
+
 * Avoid spurious byte compilation errors due to the programatically generated `spec_*()` functions.
 
 * `read_delim_chunked()` now again correctly respects the `chunk_size` parameter (#1248)